Fake nCipher classes
Enables building and running the code for people who don't own the
hardware security module. These are only placeholders for the
classes/interfaces needed by Subzero (i.e. we may add additional fakes
to this module over time).

At compile time, the fake classes provide information to the compiler.
The fake_ncipher.jar will be excluded from the shaded jar. We'll add two
entries to the manifest (one for the real jar and another for the fake
jar).

At runtime, the classloader will try loading the real jar and fallback
to the fake jar.

If the fake jar is picked, these classes don't get used thanks to the
--ncipher flag.

This all sounds convoluted and it is! We'll write some documentation
explaining how to run the code with and without actual hardware. The
instructions are going to be slightly different if you are running the code
using maven or loading the project in Intellij.