Add a server for debugging purpose.
Easiest way to run the server is:
cd java
mvn install
./server/target/server-1.0.0-SNAPSHOT.jar server

The server runs on http://localhost:8080/ and is able to render various
QR codes and process the responses.

It is not meant to be a production system. Specifically, the server
does not perform any UTXO management, does not connect to Bitcoin nodes
and does not persist any state.