Upgrade trezor-crypto
Once upon a time, trezor-firmware was composed of many smaller repos.
Trezor-crypto (which is the only piece from the Trezor project which interests
us) was a submodule. This made it easy to verify that we had not modified the
crypto layer. Luckily for us, the submodule also used cmake which made
the whole thing very easy to setup and use.

One day, trezor-crypto dropped support for cmake. Later, the various
repos were all merged into trezor-firmware. This was the last straw,
making it no longer worthwhile for us to use submodules.

This commit removes the trezor-crypto submodule and copies the crypto
folder from
https://github.com/trezor/trezor-firmware/tree/legacy/v1.8.1/crypto with
a CMakeLists.txt which we'll need to maintain ourselves.

This commit also replaces hasher_Double with HASHER_SHA2D, see
https://github.com/trezor/trezor-crypto/commit/c70e4401280fc0720922b393d71780dd2ac78611
for context.