Fix rendering bug
This temporary fix prevents typing "yes" or "no" from overwriting what's
previously on the screen. The current code is very fragile, the
rendering varies across OS platforms and many such offsets are
hardcoded.

We'll have to eventually decide if we want to completely rewrite the UI code or
if we want to refactor small pieces.