Improve codeql-analysis.yml
Build C/C++ and other languages separately, for CodeQL.
Install the latest version of protobuf from Google's github repo. This is needed because the protobuf compiler version (3.0.0) packaged with the current ubuntu-latest is not new enough for the build.
Remove git checkout HEAD^2 from the steps, as it is no longer a recommended practice.
Also make some cosmetic edits.