Differentiate ISO types: dev or release Release build: '/vagrant/build.sh' or '/vagrant/build.sh release'. Development build: '/vagrant/build.sh dev'.
A release image is to be used by cold storage operators who do not need to
have access to an interactive shell.  For a release image to boot into a
root shell successfully, an SCSI/ATA hard drive needs to be available, with
a partition '/dev/sda1' used as persistent storage for the cold wallet. The
partition can be formatted in ext3 or vfat for example. The device node
'/dev/sda1' is required to give the operators a seamless experience when
booting up the system. A few other files, including the HSM world and
encrypted secrets, and compiled subzero code (in .jar and signed .sar), are
not checked in to the project repository, and need to be added to their
appropriate locations (please refer to README.md) before building a release
image that will work out of the box.

A development image is to be used by subzero developers for HSM and cold
wallet development. A developer can boot into an interactive root shell
without performing the above mentioned steps. The HSM software is
automatically installed in the development image, but a developer needs to
manually set up persistent storage for the cold wallet, put .jar and
.sar files to appropriate filesystem locations, and execute them
accordingly, for development and testing.