SignTx regression testing for TARGET=nCipher.
The same test wallet can be used for TARGET=dev and TARGET=nCipher, in terms of SignTx regression testing. The nCipher subzero core build needs to have its associated HSM properly configured so that it shares the same AES-128-GCM keys with the off-target dev build, for master seed encryption and public key encryption. The test vectors are the same for both subzero core targets.

Note that on-target regression testing is much slower than off-target testing.