Workaround a gcc (4.8.5) bug
In theory we don't need the ctx array to be static to zero initialize it. We could have done "gcm_ctx ctx[1] = {0}" according to C standard. Unfortunately, there is a bug in a version of gcc (4.8.5) that we would like to support (cf. https://gcc.gnu.org/bugzilla/show_bug.cgi?id=53119). This change is to work around the gcc bug.