Start sshd by default for dev-flavored ISO
This change is to facilitate development. Note you may still want to run the a dev flavored ISO in a segregated network to reduce the attack surface to the development box.

This change also makes it possible/easier to build the codesafe target in CI/CD using the dev-flavored ISO.

We also improve live-usb-creator documentation.