Split AES GCM headers into dev and ncipher versions.
Because AES GCM implementations differ for the dev and nCipher targets, and M_KeyID is typedef'ed differently for both targets, we split the original aes_gcm.h into two target-specific headers, aes_gcm_dev.h and aes_gcm_ncipher.h. Previously, we use the macro STDMARSHALTYPES_H to distinguish the target type in aes_gcm.h. It requires a developer to be extra careful when including header files. This change is to make it clear to a developer which header to include in the corresponding implementation, and prevent accidential inclusions of the wrong header.

Thank oreparaz@ for bringing the potential issue to our attention.