Replace file globbing approach with explicit file list for cmake
This change makes it less difficult to produce deterministic builds, in particular, in the live ISO and in a docker container built on top of the the ISO. The problem with file globbing is that the order of globbed files is dependent on the underlying filesystem. We discovered this descrepency while building the same verson of subzero core with the same toolchain version in an identical build path, in a system booted from an live ISO (with an ext4 filesystem), and in a docker container (with an overlay filesystem) created from the same live ISO, which resulted in differed subzero binaries.