bits2qr_fb.py: tool to convert file to QR code, then output to framebuffer
We add a simple utility in the development ISO that converts an input
file into a QR code, and prints the QR code to a framebuffer. The tool
intends to facilitate data backup and transfer on unnetworked subzero
machines.

In this commit, we only add the tool to the development ISO. We may add
it to the release ISO later, if there's a need. All dependencies for
the tool have also been taken care of in this commit.

The tool is available on the live system in path
`/data/app/subzero/tools/bits2qr_fb.py`.

Example:

    echo "Testing bits2qr_fb.py" > test.txt
    /data/app/subzero/tools/bits2qr_fb.py /dev/fb0 test.txt

Synopsis:

    usage: Create QR code from file, print to framebuffer. [-h] device_node infile

    positional arguments:
      device_node  path to framebuffer device node, e.g., /dev/fb0
      infile       path to input file to convert to QR code

    optional arguments:
      -h, --help   show this help message and exit