zeroizing secrets from HSM memory. (#398)
This change defensively makes an effort to zeroize all the
possible buffers that could have been used to temporarily
store sensitive data.

The HSM provides a secure environment as such and makes it non
trivial to be able to access the RAM at runtime. But there could
be a firmware bug that might help an attacker with physical access
to an HSM to dump ram and recover secrets.

For Cold storage, public keys are also sensitive information(in order
to keep the address secret).