Adding/updating tools in the dev iso.
1. Maven version update.
2. Tooling for generating test vectors. Doc updates.

Maven version in the dev iso needs to be updated to a version
> 3.1.1 so that the build can work.

Also a lightweight x server substitute to help in running the test
vector geneartion script that does not work on mac.
Mac terminals have a buffer limit for piping and it stalls with
big enough inputs to expect scripts.

Also updating the instructions for generating the test vectors.