Include development tools only in the dev-flavored ISO
Currently, some files, including development tools and their
dependencies, are copied to the ISO during the post install step
regardless of their type (development or release). Because these files
are only useful for the dev-flavored ISO, we strip them off the
release-flavored ISO.

Also update download link to apache-maven to keep the link to
apache-maven-3.8.4-bin.tar.gz up-to-date.