Guardrail for master seed exposure
We put the code segment that prints the master seed (only for
BTC_TESTNET) behind a preprocessor macro that has to be explicitly
defined at build time, to make it less prone to mistakes. Because the
code segement is in a common file shared by the dev and nCipher targets,
it also makes code auditing easier to perform.