Fix overflow bug in conv_btc_to_satoshi()
There was an unsigned integer overflow bug in conv_btc_to_satoshi().
This didn't matter in practice because at the moment the function is
never called with large BTC inputs, but it's still a good idea to fix
it just in case this ever changes in the future.

Added a new self-check and verified that it fails w/o the change to
conv.c and passes with the change to conv.c.