[core] fix script_push() to take a uint8_t instead of enum opcodetype.
Since we can push arbitrary bytes, and enum opcodetype doesn't have a value
for every byte, let's type the function properly.

Not sure if using a non-existent enum value is technically UB or not,
but even if it's not UB it's certainly weird.

Also, enum opcode is not necessarily represented as an 8-bit integer, I think
the default representation of ints uses 32-bit integers, so there's also
potential implicit truncation issues.