[live-usb-creator] fix broken post-install script (#642)
There is a bug in the post-install script (introduced in 13bcaad), which prevents the subzero-cli.jar or the signed .sar files from being copied to the live DVD. This means it's not possible to create a working release DVD, which must contain these files. This PR fixes the issue by copying the entire data_app_subzero directory to the live DVD as was done previously.