[core] Add a fuzzing regression test github action
While fuzzing subzero core before the nanopb 0.4.5 upgrade,
the fuzzer found a bunch of inputs which triggered UBSAN.

This PR creates a regression test that runs the fuzzer with these inputs
and verifies that none of them fail now that we're on nanopb 0.4.5.

If more bad inputs are discovered later, they can be added to this directory
after the corresponding bugs are fixed.