[java] don't echo the ACS password when running the --init-nvram command
I discovered that the ACS password is echoed to the screen when running the GUI with the --init-nvram option.
This is obviously bad as anyone present in the room would be able to see the password.
Fixed by using System.console().readPassword() instead of reading from System.in.

Tested on my at-home HSM machine, both with and without a framebuffer, works as expected.