[java] refactor gradle dependencies so IntelliJ can analyze them
Due to a limitation of IntelliJ's Gradle integration, the IDE doesn't
understand gradle dependencies when they are declared in a separate file
and then referenced from build.gradle.kts. However, if only the dependency
version is declared in another file and string interpolation is used, then
the IDE can see the dependencies without any issue. This change refactors
the gradle dependencies to make the IntelliJ integration work.

This is a prerequsite for detecting vulnerable dependencies with IntelliJ
and upgrading them.